+++ src/cli/mod.rs
@@ -1,0 +1,6 @@
+mod args;
+mod color_mode;
+mod help;
+mod output;
+
+pub use self::{args::*, color_mode::*, help::*, output::*};