+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Select where the results will be written.
 - Improve errors.
 - Option to disable color.
+- Option to capture output.
 
 ## v0.1.1
 - Measure proces run time and peak memory.